requirements.txt
colorama==0.4.6 
tabulate==0.9.0
iso3166==2.1.1
customtkinter==0.3
pillow==11.2.1
PySide==6.9.0